refactor: move as callable function for scripting usage
